refactor: move object definitions to their own modules
Move Git object definitions to their own modules, alongside their
instances for the `Object` typeclass.
